functional recycler and details fragment-no images

+++ app/src/main/java/com/myapp/walkme/ui/fragments/OnDogSelectedListener.kt
@@ -1,5 +1,5 @@
 package com.myapp.walkme.ui.fragments
 
 interface OnDogSelectedListener {
-    fun onDogSelected(position: Int)
+    fun onDogSelected(position: Long)
 }
